requirements.txt
elevenlabs
pyaudio
openai
python-dotenv
pillow
langchain_community
streamlit
gradio
watchdog